+++ src/services/review.ts
@@ -1,4 +1,4 @@
-import { UserReviewForm } from '@typings/review';
+import { UserReviewEditForm, UserReviewForm } from '@typings/review';
 import { UserReviewsResponse } from '@typings/response/review';
 import { fetchCall } from './api';
 
@@ -13,4 +13,15 @@ export const postNewReview = async (
 export const getUserReviews = async (cursor?: number) => {
   const url = cursor ? `users/reviews?cursor=${cursor}` : 'users/reviews';
   return await fetchCall<UserReviewsResponse>(url, 'get');
+};
+
+export const deleteReview = async (reviewId: string) => {
+  return await fetchCall(`users/reviews/${reviewId}`, 'delete');
+};
+
+export const editReview = async (
+  reviewId: string,
+  editedReview: UserReviewEditForm,
+) => {
+  return await fetchCall(`users/reviews/${reviewId}`, 'put', editedReview);
 };
